Enable bugprone-misplaced-widening-cast
GetExperienceRequiredForLevel computed 1000 * effectiveLevel in int and
only widened the product to uint64_t afterwards, so the multiplication
overflowed from level 2072 up. The experience tooltip walks its level
variable as far as 10000, so the argument does reach that range. Widen
before multiplying, then drop the exclusion.

+++ src/GUI/UI/UIPopup.cpp
@@ -2335,5 +2335,5 @@ uint64_t GetExperienceRequiredForLevel(int level) {
     int effectiveLevel = 0;
     for (int i = 0; i < level; ++i)
         effectiveLevel += i + 1;
-    return (uint64_t)(1000 * effectiveLevel);
+    return 1000ULL * effectiveLevel;
 }
